Change TO DS Versioning to be self-contained
Changes TO DS Versioning, so new versions don't have to change
existing version code. Only the new version has to be added.

Further, the new version is limited to the inevitable struct
and sanitize/convert funcs, plus a handler which can be
copied and sed-replaced (and replaced with Generics in Go2)