Add Edge MSO to Perl Cache Config
Perl is no longer used in master; but, in the event it's somehow
accidentally called, this will prevent it from doing the wrong then.

It's an extra safety, and there's really no reason not to.

It also provides a changeset to apply it, which anyone using an older
Traffic Ops can backport if they need.