Add ORT client-side update_statuses generation
This adds a flag to the ORT atstccfg helper app, to generate the
update_status, primarily parent pending data, from standard TO
data API endpoints.

The advantage of this is moving more logic client-side, which can be
canaried, more easily deployed and rolled back, etc than the Traffic
Ops server.

The disadvantage is that it takes considerably longer to get all data,
especially for a large CDN using many DeliveryServiceServers and not
using Topologies.

For now, this only adds the logic and flag to the helper, but doesn't
make ORT use it.

Once Topologies are more common, or ORT/t3c is changed to be more
modular, those disadvantages go away and we should make ORT use the
client-side logic.